# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Niu Gateway API
  version: 0.1.0
  description: Public inference, health, and initial management endpoints for Niu.
servers:
  - url: /
paths:
...
  /v1/models:
    get:
      operationId: listModels
      summary: List models available to the configured API key
      security:
        - bearerAuth: []
        - niuApiKeyAuth: []
      responses:
        '200':
          description: Available public model names.
        '401':
          description: Missing or invalid gateway credentials.
...
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      summary: Create an OpenAI-compatible chat completion
      security:
        - bearerAuth: []
        - niuApiKeyAuth: []
      parameters:
        - name: X-Niu-Task-ID
          in: header
          required: false
          description: Optional opaque task correlation key. Requests with the same value can be grouped in project activity. Niu stores the value as metadata and does not forward it to the provider.
          schema: { type: string, minLength: 1, maxLength: 200, pattern: '^[!-~]{1,200}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model:
                  type: string
                messages:
                  type: array
                  items:
                    type: object
                tools:
                  type: array
                  minItems: 1
                  maxItems: 128
                  items:
                    $ref: '#/components/schemas/ChatFunctionTool'
                tool_choice:
                  $ref: '#/components/schemas/ChatToolChoice'
                response_format:
                  $ref: '#/components/schemas/ChatResponseFormat'
              additionalProperties: true
      responses:
        '200':
          description: A normalized chat completion or compatible event stream.
        '400':
          description: Invalid request or unsupported operation.
        '401':
          description: Missing or invalid gateway credentials.
        '501':
          description: The requested tool or structured-output capability is disabled for this route.
        '502':
          description: Provider request failed.
      x-niu-status: Basic chat is implemented for configured native provider routes. Function tools, streaming tool deltas, and structured JSON are opt-in on OpenAI-compatible routes only. Niu validates tool-call shape and JSON output, but does not execute tools or validate JSON Schema semantics. Structured JSON streaming and priced tool calls are unsupported.
  /v1/embeddings:
    post:
      operationId: createEmbedding
      summary: Create OpenAI-compatible text embeddings
      description: The configured route must use provider openai and explicitly declare embedding support. Optional dimensions and base64 output require separate route capabilities. These declarations are operator assertions, not provider conformance evidence. Unsupported providers or capabilities are rejected before durable operation and attempt creation.
      security:
        - bearerAuth: []
        - niuApiKeyAuth: []
      parameters:
        - name: X-Niu-Task-ID
          in: header
          required: false
          description: Optional opaque task correlation key. Requests with the same value can be grouped in project activity. Niu stores the value as metadata and does not forward it to the provider.
          schema: { type: string, minLength: 1, maxLength: 200, pattern: '^[!-~]{1,200}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: One validated embedding per input item; the configured upstream model is replaced with the public model alias.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '400':
          description: Invalid body, unsupported field, or invalid input shape.
        '413':
          description: Request body exceeds the gateway's configured body limit.
        '401':
          description: Missing or invalid gateway credentials.
        '404':
          description: Model alias does not exist or is unavailable to this key.
        '501':
          description: Configured provider or declared route capability does not support this operation.
        '502':
          description: Provider request failed or returned an invalid embedding response.
        '422':
          description: Request body is not valid JSON.
      x-niu-status: OpenAI-compatible text inputs and float/base64 response vectors; no token-ID arrays, multimodal inputs, or streaming.
  /v1/responses:
    post:
      operationId: createResponse
      summary: Create a non-streaming text response
      description: The route must be an OpenAI-compatible provider route with supports_responses enabled. This public subset accepts a single text input and optional text instructions, output limit, sampling values, metadata and user identifier. Streaming, multimodal input, tools, prior-response state and other fields are rejected.
      security:
        - bearerAuth: []
        - niuApiKeyAuth: []
      parameters:
        - name: X-Niu-Task-ID
          in: header
          required: false
          description: Optional opaque task correlation key. Requests with the same value can be grouped in project activity. Niu stores the value as metadata and does not forward it to the provider.
          schema: { type: string, minLength: 1, maxLength: 200, pattern: '^[!-~]{1,200}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsesRequest'
      responses:
        '200':
          description: A validated Responses API text response with the public model alias.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsesResponse'
        '400':
          description: Invalid body, unsupported field or invalid request shape.
        '401':
          description: Missing or invalid gateway credentials.
        '404':
          description: Model alias does not exist or is unavailable to this key.
        '501':
          description: Responses support is disabled for the route or streaming was requested.
        '502':
          description: Provider request failed or returned an invalid text response.
      x-niu-status: Non-streaming, text-only subset on explicitly configured OpenAI-compatible routes. The gateway validates response item types and usage shape. Tool, multimodal, streaming and stateful conversation semantics are not implemented.
  /admin/v1/models:
    get:
      operationId: adminListModels
      summary: List configured model routes
      security:
        - adminBearerAuth: []
      responses:
        '200':
          description: Public and upstream model names with provider identifiers and catalog visibility.
        '401':
          description: Missing or invalid administrator credentials.
  /admin/v1/vendors:
    get:
      operationId: listVendors
      summary: List installation vendor metadata without credentials
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Vendor metadata. All management responses use Cache-Control no-store.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Vendor' } }
        '403': { description: Installation administrator required. }
    post:
      operationId: createVendor
      summary: Create an installation vendor with an encrypted credential
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorCreate' }
      responses:
        '201': { description: Created vendor metadata wrapped in data. The credential is never returned. }
        '400': { description: Invalid vendor configuration. }
        '403': { description: Installation administrator required. }
        '409': { description: Vendor name already exists. }
        '503': { description: Storage or credential encryption unavailable. }
  /admin/v1/vendors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: updateVendor
      summary: Update metadata or rotate a vendor credential
      description: Adapter is immutable. Changes affect new requests; already dispatched work can finish using its original configuration.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorUpdate' }
      responses:
        '200': { description: Updated vendor metadata wrapped in data. }
        '400': { description: Invalid configuration or credential. }
        '403': { description: Installation administrator required. }
        '404': { description: Vendor does not exist. }
        '409': { description: Stale revision or conflicting name. }
  /admin/v1/vendors/{id}/models:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listVendorModels
      summary: List model mappings for an installation vendor
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Model mappings wrapped in data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/VendorModel' } }
        '403': { description: Installation administrator required. }
        '404': { description: Vendor does not exist. }
    post:
      operationId: upsertVendorModel
      summary: Create or update a model mapping with optimistic revision checks
      description: Omit expected_revision to create; provide the current revision to update. Aliases cannot be reassigned to another vendor. Disabled mappings still override static file aliases.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorModelInput' }
      responses:
        '200': { description: Persisted model mapping wrapped in data. }
        '400': { description: Invalid alias, capabilities or pricing. }
        '403': { description: Installation administrator required. }
        '404': { description: Vendor does not exist. }
        '409': { description: Stale revision, existing alias or different vendor ownership. }
  /admin/v1/vendors/{id}/catalog:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listProviderModelCatalog
      summary: Discover models available from an installation vendor
      description: Makes a bounded GET to the configured provider /models endpoint using the encrypted server-side credential. No inference request is sent. Redirects are blocked, the response body is capped at 2 MiB, and only model ID, display name, and context length are returned. Provider credentials and raw response data are never returned.
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Provider model metadata wrapped in data.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    maxItems: 2000
                    items: { $ref: '#/components/schemas/ProviderCatalogModel' }
        '403': { description: Installation administrator required. }
        '404': { description: Vendor does not exist. }
        '502': { description: Provider rejected the credential or returned an invalid catalog or request error. }
  /admin/v1/vendors/{id}/check:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: checkVendorModel
      summary: Check provider reachability and whether a mapped model is listed
      description: Performs a bounded GET to the provider's /models endpoint. It does not send an inference request. Redirects are blocked, response bodies are capped at 2 MiB, and provider response content and credentials are never returned. A listed model does not prove inference entitlement or quota.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [alias]
              properties:
                alias: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '200':
          description: Sanitized provider check result wrapped in data.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status, model, http_status, duration_ms, checked_at_ms]
                    properties:
                      status:
                        type: string
                        enum: [connected, credentials_rejected, endpoint_unavailable, private_endpoint_blocked, invalid_endpoint, redirect_blocked, provider_rate_limited, provider_error, model_catalog_unavailable, model_catalog_too_large, invalid_model_catalog]
                      model: { type: string, enum: [listed, not_listed, unknown] }
                      http_status: { type: [integer, 'null'] }
                      duration_ms: { type: integer, minimum: 0 }
                      checked_at_ms: { type: integer, minimum: 0 }
        '400': { description: Invalid model alias. }
        '403': { description: Installation administrator required. }
        '404': { description: Model alias is not mapped to this vendor. }
        '503': { description: Credential decryption or storage unavailable. }
  /admin/v1/session:
    get:
      operationId: getAdminSessionIdentity
      summary: Read the current administrator's scope and permissions
      description: Revalidates the bearer credential on every request. Returns installation authority or the operator's current organization/project scope and role. Permissions are UI hints; every management operation independently enforces authorization. Responses contain no credentials and use Cache-Control no-store.
      security:
        - adminBearerAuth: []
      responses:
        '200':
          description: Current identity and effective role permissions.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [kind, operator, permissions]
                    properties:
                      kind: { type: string, enum: [installation, operator] }
                      operator:
                        type: [object, 'null']
                        required: [id, role, organization_id, project_id]
                        properties:
                          id: { type: string, format: uuid }
                          role: { type: string, enum: [owner, admin, viewer] }
                          organization_id: { type: string, format: uuid }
                          project_id: { type: [string, 'null'], format: uuid }
                      permissions:
                        type: object
                        required: [read, write, manage_operators]
                        properties:
                          read: { type: boolean }
                          write: { type: boolean }
                          manage_operators: { type: boolean }
        '401':
          description: Missing, invalid, expired or revoked administrator credential.
        '503':
          description: Operator storage is unavailable.
  /admin/v1/operators:
    get:
      operationId: adminListOperators
      summary: List operator identities without session credentials
      description: Requires a workspace owner or a configured bootstrap administrator token. Results are limited to the caller's organization/project unless the caller uses installation bootstrap credentials. Responses contain session identifiers and expiry metadata but never token hashes or bearer tokens.
      security:
        - adminBearerAuth: []
      responses:
        '200':
          description: Persistent operator identities and assigned roles.
        '401':
          description: Missing, invalid, expired or revoked credentials.
        '403':
          description: The authenticated operator is not an owner.
    post:
      operationId: adminCreateOperator
      summary: Create an operator and issue its first session token
      description: Requires a workspace owner or a configured bootstrap administrator token. Every operator must be bound to an organization and may optionally be limited to one project. The bearer token is returned once and is never retrievable from list APIs.
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorInput'
      responses:
        '201':
          description: Operator and session created; response includes Cache-Control no-store.
        '400':
          description: Invalid role, name or session lifetime.
        '401':
          description: Missing, invalid, expired or revoked credentials.
        '403':
          description: The authenticated operator is not an owner.
  /admin/v1/operators/{operator}:
    delete:
      operationId: adminRevokeOperator
      summary: Revoke an operator and all of its sessions
      security:
        - adminBearerAuth: []
      parameters:
        - name: operator
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Operator and active sessions revoked.
        '401':
          description: Missing, invalid, expired or revoked credentials.
        '403':
          description: The authenticated operator is not an owner.
        '409':
          description: Operator was already revoked or does not exist.
  /admin/v1/operators/{operator}/sessions:
    get:
      operationId: adminListOperatorSessions
      summary: List session expiry and revocation metadata
      security:
        - adminBearerAuth: []
      parameters:
        - name: operator
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Session identifiers and expiry metadata; credentials are never returned.
        '401':
          description: Missing, invalid, expired or revoked credentials.
        '403':
          description: The authenticated operator is not an owner.
    post:
      operationId: adminIssueOperatorSession
      summary: Issue an additional one-time operator session token
      security:
        - adminBearerAuth: []
      parameters:
        - name: operator
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorSessionInput'
      responses:
        '201':
          description: Session created; response includes Cache-Control no-store and the token is shown once.
        '400':
          description: Session lifetime is outside the supported range.
        '401':
          description: Missing, invalid, expired or revoked credentials.
        '403':
          description: The authenticated operator is not an owner.
        '409':
          description: Operator is unavailable.
  /admin/v1/operators/{operator}/sessions/{session}:
    delete:
      operationId: adminRevokeOperatorSession
      summary: Revoke one operator session
      security:
        - adminBearerAuth: []
      parameters:
        - name: operator
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: session
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Session revoked.
        '401':
          description: Missing, invalid, expired or revoked credentials.
        '403':
          description: The authenticated operator is not an owner.
        '409':
          description: Session was already revoked or does not exist.
...
components:
  schemas:
...
    OperatorSessionInput:
      type: object
      additionalProperties: false
      required: [expires_in_seconds]
      properties:
        expires_in_seconds: { type: integer, minimum: 60, maximum: 31536000 }
...
    ChatFunctionTool:
      type: object
      required: [type, function]
      properties:
        type: { type: string, const: function }
        function:
          type: object
          required: [name]
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 64
              pattern: '^[A-Za-z0-9_-]+$'
            description: { type: string }
            parameters:
              type: object
              description: JSON Schema object forwarded to the configured provider.
              additionalProperties: true
            strict: { type: boolean }
          additionalProperties: true
      additionalProperties: true
    ChatToolChoice:
      oneOf:
        - type: string
          enum: [none, auto, required]
        - type: object
          required: [type, function]
          properties:
            type: { type: string, const: function }
            function:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 64 }
          additionalProperties: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    niuApiKeyAuth:
      type: apiKey
      in: header
      name: X-Niu-API-Key
    adminBearerAuth:
      type: http
      scheme: bearer